+++ pycvvdp/run_cvvdp.py
@@ -94,6 +94,7 @@ def parse_args(arg_list=None):
     parser.add_argument("--pix-per-deg", type=float, default=None, help='Overwrite display geometry and use the provided pixels per degree value.')
     parser.add_argument("--fps", type=float, default=None, help='Frames per second. It will overwrite frame rate stores in the video file. Required when passing an array of image files.')
     parser.add_argument("--frames", type=str, default=None, help='Range of frames specified as first:step:last, first:last, or first: (Matlab notation). Currently works only with frames provided as images.')
+    parser.add_argument("--gpu-mem", type=float, default=None, help='How much GPU memory can we use in GB. Use if CUDA reports out of mem errors, or you want to run multiple instances at the same time.')
     parser.add_argument("-q", "--quiet", action='store_true', default=False, help="Do not print any information but the final JOD value. Warning message will be still printed.")
     parser.add_argument("-v", "--verbose", action='store_true', default=False, help="Print out extra information.")
     parser.add_argument("--ffmpeg-cc", action='store_true', default=False, help="Use ffmpeg for upsampling and colour conversion. Use custom pytorch code by default (faster and less memory).")
@@ -231,6 +232,7 @@ def run_on_args(args):
                                 temp_padding=args.temp_padding,
                                 config_paths=args.config_paths,
                                 quiet=args.quiet,
+                                gpu_mem=args.gpu_mem,
                                 dump_channels=dump_channels )
             metrics.append( fv )
         elif mm == 'pu-psnr-rgb':